+++ srcs/parse_utils/search_env.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   search_env.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yubchoi <yubchoi@student.42.fr>            +#+  +:+       +#+        */
+/*   By: hako <hako@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/07/21 19:42:09 by yubchoi           #+#    #+#             */
-/*   Updated: 2022/07/21 19:42:09 by yubchoi          ###   ########.fr       */
+/*   Created: 2022/07/22 22:24:41 by hako              #+#    #+#             */
+/*   Updated: 2022/07/22 22:24:49 by hako             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
